Passing unit tests for auth.py
